add confirmation when deleting stuff, add product categories, navbar changes

+++ client/src/components/AdminPanel/Coupons/SingleCoupon.js
@@ -60,8 +60,10 @@ const SingleCoupon = ({ setCurrent }) => {
   };
 
   const remove = () => {
-    dispatch(removeCoupon(id));
-    history.push('/admin/coupons');
+    if (window.confirm('Are you sure you want to remove this coupon?')) {
+      dispatch(removeCoupon(id));
+      history.push('/admin/coupons');
+    }
   };
 
   if (coupon.failed) {
